% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyr_lang_detect.R
\name{cyr_lang_detect}
\alias{cyr_lang_detect}
\title{Determine the language of a Cyrillic text}
\usage{
cyr_lang_detect(text)
}
\arguments{
\item{text}{A string of Cyrillic characters, between " ", may contain spaces.}
}
\value{
A message with the language of the input text.
}
\description{
Determine the language of a Cyrillic text
}
\examples{
cyr_lang_detect("Лише раз на сто років Небо народжує двох людей з однією душею
І відпускає їх у чистий, вільний політ Нам пощастило Я люблю тебе")
}
