% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eng_ukr_transliterate.R
\name{eng_ukr_transliterate}
\alias{eng_ukr_transliterate}
\title{Transliterate a text from English Latin to Ukrainian Cyrillic script}
\usage{
eng_ukr_transliterate(text)
}
\arguments{
\item{text}{A string of English Latin characters, between " ", may contain spaces.}
}
\value{
A string of characters transliterated to Ukrainian Cyrillic script.
}
\description{
Transliterate a text from English Latin to Ukrainian Cyrillic script
}
\examples{
eng_ukr_transliterate("Sail and seek The starbound quay
Calling you, calling me To be a part of your story")
}
