% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ukrainian.R
\name{is_ukrainian}
\alias{is_ukrainian}
\title{Determine if a text is Ukrainian}
\usage{
is_ukrainian(text)
}
\arguments{
\item{text}{A string of Cyrillic characters, between " ", may contain spaces.}
}
\value{
A message if the input text is Ukrainian or not.
}
\description{
Determine if a text is Ukrainian
}
\examples{
is_ukrainian("Лише раз на сто років Небо народжує двох людей з однією душею
І відпускає їх у чистий, вільний політ Нам пощастило Я люблю тебе")
}
