% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukr_eng_transliterate.R
\name{ukr_eng_transliterate}
\alias{ukr_eng_transliterate}
\title{Transliterate a text from Ukrainian Cyrillic to English Latin script}
\usage{
ukr_eng_transliterate(text)
}
\arguments{
\item{text}{A string of Ukrainian Cyrillic characters, between " ", may contain spaces.}
}
\value{
A string of characters transliterated to English Latin script.
}
\description{
Transliterate a text from Ukrainian Cyrillic to English Latin script
}
\examples{
ukr_eng_transliterate("Лише раз на сто років Небо народжує двох людей з однією душею
І відпускає їх у чистий, вільний політ Нам пощастило Я люблю тебе")
}
